pyproject.toml
[tool.poetry]
name = "teinou-bot"
version = "0.0.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "teinou_bot"}]

[tool.poetry.dependencies]
python = "^3.11"
python-dotenv = "^1.0.0"
discord-py = "^2.3.2"

[tool.poetry.scripts]
start = "src:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
